requirements.txt
Django==3.2.9
psycopg2-binary==2.9.2
gunicorn==20.1.0
djangorestframework==3.12.4
django-filter==21.1
